Use templates for math classes

+++ sr/Constants.hpp
@@ -7,8 +7,8 @@
 
 namespace sr
 {
-    constexpr float tau = 6.28318530717958647692F;
-    constexpr float pi = 3.14159265358979323846F;
+    constexpr double tau = 6.28318530717958647692;
+    constexpr double pi = 3.14159265358979323846;
 }
 
 #endif
